東京23区新規登録・更新・削除機能追加（竹原）

+++ web-app-training/src/main/java/jp/co/futureantiques/webapptraining/model/tokyoTakehara/TokyoMainTakehara.java
@@ -45,7 +45,7 @@ public class TokyoMainTakehara {
 
 	/** 面積(㎢) */
 	@Column(name = "area")
-	private Integer area;
+	private Float area;
 
 	/** 区役所の最寄駅 */
 	@Column(name = "ward_office")
